(wip) created mixin plugin and migrated templates to new directories
