+++ ckanext/googleanalyticsbasic/templates/templates_2_8/base.html
