+++ ckanext/googleanalyticsbasic/templates/templates_2_8/snippets/googleanalyticsbasic_header.html
